Refactor CellView
 * Join CellViewLeft/CellViewRight into CellView
 * Remove unnecessary cell view xml layouts
