+++ app/src/main/res/layout/view_cell.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:minHeight="44dp"
+    tools:layout_height="60dp"
     android:background="@drawable/clickable_background_color_lawrence">
 
     <View
@@ -15,17 +17,138 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <io.horizontalsystems.bankwallet.components.CellLeftView
-        android:id="@+id/cellLeft"
-        android:layout_width="0dp"
-        android:layout_height="match_parent"
-        app:layout_constraintEnd_toStartOf="@+id/cellRight"
-        app:layout_constraintStart_toStartOf="parent" />
+    <!-- Cell Left -->
 
-    <io.horizontalsystems.bankwallet.components.CellRightView
-        android:id="@+id/cellRight"
+    <io.horizontalsystems.bankwallet.ui.extensions.IconView
+        android:id="@+id/cellIcon"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/ic_attention" />
+
+    <TextView
+        android:id="@+id/cellTitle"
+        style="@style/Body"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:textColor="?TextColorBarsToDark"
+        app:layout_constraintVertical_chainStyle="packed"
+        app:layout_constraintBottom_toTopOf="@+id/cellSubtitle"
+        app:layout_constraintStart_toEndOf="@+id/cellIcon"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="Title" />
+
+    <TextView
+        android:id="@+id/cellSubtitle"
+        style="@style/Subhead2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textColor="@color/grey"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="@+id/cellTitle"
+        app:layout_constraintTop_toBottomOf="@+id/cellTitle"
+        tools:text="Subtitle" />
+
+    <TextView
+        android:id="@+id/coinTypeLabel"
+        style="@style/TextLabel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="4dp"
+        android:visibility="gone"
+        app:layout_constraintBottom_toBottomOf="@+id/cellSubtitle"
+        app:layout_constraintStart_toEndOf="@+id/cellSubtitle"
+        app:layout_constraintTop_toTopOf="@+id/cellSubtitle"
+        tools:text="ERC20"
+        tools:visibility="visible"/>
+
+    <!-- Cell Right -->
+
+    <TextView
+        android:id="@+id/cellLabel"
+        style="@style/Body"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
+        android:gravity="center"
+        android:paddingStart="0dp"
+        android:paddingEnd="16dp"
+        android:textColor="@color/grey"
+        app:layout_constraintEnd_toStartOf="@id/lightModeIcon"
+        tools:text="Label" />
+
+    <ImageView
+        android:id="@+id/lightModeIcon"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:tint="?ColorJacob"
+        app:layout_constraintEnd_toStartOf="@+id/dropDownIcon"
+        app:srcCompat="@drawable/ic_light_mode" />
+
+    <TextView
+        android:id="@+id/dropdownValue"
+        style="@style/Body"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:textColor="?TextColorBarsToDark"
+        android:visibility="gone"
+        app:layout_constraintEnd_toStartOf="@+id/dropDownIcon"
+        tools:text="Label" />
+
+    <ImageView
+        android:id="@+id/dropDownIcon"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="16dp"
+        android:tint="@color/grey"
+        app:layout_constraintEnd_toStartOf="@+id/arrowIcon"
+        app:srcCompat="@drawable/ic_down" />
+
+    <ImageView
+        android:id="@+id/badgeImage"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:src="@drawable/ic_attention"
+        android:tint="?ColorLucian"
+        android:visibility="gone"
+        app:layout_constraintEnd_toStartOf="@+id/arrowIcon" />
+
+    <ImageView
+        android:id="@+id/arrowIcon"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:tint="@color/grey"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:srcCompat="@drawable/ic_arrow_right" />
+
+    <ImageView
+        android:id="@+id/checkIcon"
+        android:layout_width="18dp"
+        android:layout_height="match_parent"
+        android:layout_marginStart="12dp"
+        android:layout_marginEnd="16dp"
+        android:tint="?ColorJacob"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:srcCompat="@drawable/ic_checkmark" />
+
+    <Switch
+        android:id="@+id/switchView"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_marginEnd="16dp"
+        android:checked="true"
+        android:saveEnabled="false"
+        android:theme="@style/SwitchTheme"
+        android:visibility="gone"
         app:layout_constraintEnd_toEndOf="parent" />
 
     <View
