--- app/src/main/res/layout/view_cell_left.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <io.horizontalsystems.bankwallet.ui.extensions.IconView
-        android:id="@+id/cellIcon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/ic_attention" />
-
-    <TextView
-        android:id="@+id/cellTitle"
-        style="@style/Body"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:textColor="?TextColorBarsToDark"
-        app:layout_constraintVertical_chainStyle="packed"
-        app:layout_constraintBottom_toTopOf="@+id/cellSubtitle"
-        app:layout_constraintStart_toEndOf="@+id/cellIcon"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="Title" />
-
-    <TextView
-        android:id="@+id/cellSubtitle"
-        style="@style/Subhead2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/grey"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="@+id/cellTitle"
-        app:layout_constraintTop_toBottomOf="@+id/cellTitle"
-        tools:text="Subtitle" />
-
-    <TextView
-        android:id="@+id/coinTypeLabel"
-        style="@style/TextLabel"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="4dp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="@+id/cellSubtitle"
-        app:layout_constraintStart_toEndOf="@+id/cellSubtitle"
-        app:layout_constraintTop_toTopOf="@+id/cellSubtitle"
-        tools:text="ERC20"
-        tools:visibility="visible"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
